Update signup.php
I've been facing some syntax errors in the tables. One of them is working fine, while the other isn't. If I am using ".." for variables in the tables for the entries, the entire signup page is not being displayed. https://stackoverflow.com/questions/38502212/how-to-use-php-variables-as-table-name-to-insert-records-in-mysql,. I followed this syntax then but still things aren't working nicely.